[MIRROR] [Ready]Adds direct pay for cargo orders (#529)
* [Ready]Adds direct pay for cargo orders (#40684)

cl XDTM
add: You can now pay for cargo orders from your account with the cargo requests console. Credits will be detracted from the requester's account instead of the cargo budget. A 10% handling fee on top of the order will be paid to the cargo department budget.
add: Cargo employees must still accept the order for it to be delivered.
add: The delivery will arrive in a locked crate that can only be opened by an id with the paying bank account.
/cl

If you're wealthy, why should you waste your time waiting for cargo to get money? Pay for the stuff yourself; the quartermaster won't have to worry about their budget and will likely just put the order in the next shipment.

* [Ready]Adds direct pay for cargo orders