[MIRROR] Security Webbing (#546)
* Security Webbing (#41394)

This PR sees the addition of an expensive item to the Security vendor, the security webbing. A unique chest rig.

The concept of this item is to give Officers something to spend their money on when they save up.

webbing

The chest rig is intended as an expensive minor upgrade over the standard security belt that security personal can purchase. Functionally it is identical to the security belt except it can hold one extra item, up to 6 from 5.

The belt costs $800, meaning round start Officers will be unable to afford it. The sec vendor holds five of these webbings.

cl Steelpoint
add: Nanotrasen Security has unveiled a newly designed Security Webbing, a minor upgrade over the standard issue security belt that is comfortable, tactical and able to hold an extra security item.
add: However, the webbing is not considered standard issue, so any interested officers will need to buy the item from a SecVendor out of their own pay.
/cl

* Security Webbing