[MIRROR] Reduces the amount of antags spawned by the Summon Guns and Summon Magic rituals (#559)
* reduces chance from 25 to 10 for both types of rightandwrong (#41151)

cl
balance: The Summon Guns and Summon Magic rituals will now only turn 10% of the crew each into antagonists, down from 25%.
/cl

This has frustrated people for a long time. The issue is that these rituals irrevocably destroy the round that they're in by turning one dangerous murderboning antag into literally half the station becoming dangerous murderboning antags. I reduced it to 10% each. I wanted to reduce it to zero so that they wouldn't delay the round for 30 minutes after the wizard teleports in and immediately dies, but I figured that simply reducing the amount of antags spawned would be better received.

* Reduces the amount of antags spawned by the Summon Guns and Summon Magic rituals