Paystand Modifications (#41588)
cl MMMiracles
add: Signalers can now be attached to active paystands to send a signal when a certain amount of money is deposited.
add: Swiping your card on a paystand you own will let you lock it down, preventing it from being unbolted from the ground.
tweak: Paystands now ask for how much you want to deposit when interacting instead of a static price.
/cl

I thought it was pretty dumb the paystands designed to encourage shops was also forced to a single static price instead of allowing the person to put in how much they actually wanted to pay.

The signaler attaching lets you hook up a signaler to any active paystand and set up a threshold. Any amount of money deposited after this will start counting up to the threshold and reset when said threshold is hit, activating the signaler in the process.