[MIRROR] Gibs now rot and generate miasma (#794)
* Gibs now rot and generate miasma (#41782)

* Gibs now rot and generate miasma

:cl: coiax
add: Gibs will now rot if not cleaned, and produce small amounts of miasma, approximately
equal to a quarter of a corpse.
/:cl:

Miasma is fun. Gives the janitor more of a reason to do work. It will only become
dangerous if you have a lot of gibs in a small space. So better keep those Hopline shutters
open.

* Rot component, for things that rot

* Whoops

* No longer processes, so doesn't need to stop

* Gibs now rot and generate miasma