[MIRROR] Adds Paraplegic quirk (FREE ROUNDSTART WHEELCHAIR*) (#1160)
* Adds Paraplegic quirk (#42503)

:cl:
add: Adds the Paraplegic quirk for -3 points. You start with unhealable leg paralysis
(persists through cloning), and have a wheelchair to move around the station.
/:cl:

This is really popular for some reason.

This required a surprising amount of small code tweaks for it to "feel"
right in edge cases like being a job that had items in their hands or
joining on the arrivals shuttle.

Justification for -3 points: You move REALLY slowly, in a game that's
about generally running from security/the guy with the esword. And if
you die and get cloned and don't have your chair, you move even slower.

* Adds Paraplegic quirk (FREE ROUNDSTART WHEELCHAIR*)