Merge pull request #42899 from MrStonedOne/sqlstickybans
Stickybans are stored in the database.

Because byond has to keep its own list, most of this is about ferrying info between the database and byond, and keeping them synced up. Because of this, this system will properly convert existing file stickybans to database stickybans during first init. This system also plays nicely with file stickybans existing on both servers (ish): seen key matches will be merged together in the database, but the reason, date, and admin, will be taken from the first server to introduce the stickyban to the database.

This has further protections from rogue bans, but byond fixed those. I'm keeping them in anyways because they will be helpful if this ever becomes an issue again. They are: Enforcement of matches on already connected users are delayed for 5 seconds to see if the match will turnout to be fucked. Confirmed fucked stickybans will be disabled for the round.

This fixes a bug that kept stickybans from properly doing hub backed matching on servers that were not currently on the hub.

This adds a button to whitelist a ckey from a given stickyban, this uses an internal and undocumented byond stickyban field.

This adds a button to allow admins to disable the stickyban for the current server for the current round. Called putting the ban on "Time out".