[MIRROR] Gives space cleaner the ability to clean people (#1644)
* Gives space cleaner the ability to clean people

* Give space cleaner the ability to clean people (#43494)

Space cleaner sprayed, splashed or vaporised onto someone will slightly
improve their hygiene. Effective janitors can spray stinky spessmen to
stave off the stench long enough for the offender to have a shower. Space
cleaner will only stop you from stinking for a short time, unlike a shower
which will keep you fresh for much longer.

The amount of cleanliness from a single application of space cleaner is
capped, so a chemist dumping a bluespace beaker of space cleaner on their
head is not going to be as effective as a shower.

Giving janitors an extra way to help the crew is probably good, especially
because this additional mechanic can't be replaced by janiborgs or
cleanbots.