[MIRROR] Fixes movement hunger costs. (#1685)
* Fixes movement hunger costs.

* Fix movement hunger costs (#43607)

It seems it was intended to make movement cost a bit of hunger per tile,
but was actually only happening while you were floating (despite the
comment saying it should be easy).

This reverses that so it's now working as intended.

Testing shows this increases nutrition by about 5.6 per minute of
non-stop (non-floating) running movement. Walking cuts this cost by more
than half as not only do you cover fewer tiles in a minute, but walking
also costs half as much in tile-tile movement.

If you're on meth, expect to get quite hungry from zooming around the
station like a madman at 300 mph, which I think makes quite a bit of
sense.